Add base test case for inheritance for other test clases]
