+++ tests/test_group.py
@@ -1,39 +1,7 @@
-import unittest
+from tests.base import BaseTest
 import json
-from api.app import create_app
-from database.db import DatabaseConnection
-from api.utilitiez.auth_token import encode_token
-from api.models.user import User
-from api.models.group import Group
 
-
-class MessageTestCase(unittest.TestCase):
-
-    def setUp(self):
-        """initializing method for a unit test"""
-        self.app = create_app("Testing")
-        self.client = self.app.test_client(self)
-        self.db = DatabaseConnection()
-        self.user_data = {
-            "firstname": "Bekalaze",
-            "lastname": "Joseph",
-            "email": "bekeplar@gmail.com",
-            "password": "Bekeplar1234"
-                }
-            
-        self.user_login_data = {
-                       "email":"bekeplar@gmail.com",
-                       "password": "Bekeplar1234"
-                     }
-
-        self.group_data = {
-            "group_name": "Andela21",
-        }
-
-        self.user_id = 1
-        self.token = encode_token(self.user_id)
-        self.data = {}
-
+class MessageTestCase(BaseTest):
 
     def test_create_new_group(self):
         self.client.post('/api/v1/auth/signup', content_type="application/json", data=json.dumps(self.user_data))        
@@ -156,7 +124,5 @@ class MessageTestCase(unittest.TestCase):
         self.assertIsInstance(response_data, dict)
 
 
-    def tearDown(self):
-        self.db.drop_table('users')
-        self.db.drop_table('groups')
+
